pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gao-dev"
version = "1.0.0"
description = "GAO-Dev - Software Engineering Team for Generative Autonomous Organisation"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "GAO Development Team"}
]
keywords = ["ai", "agents", "autonomous", "development", "gao"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pyyaml>=6.0",
    "gitpython>=3.1.0",
    "click>=8.1.0",
    "anthropic>=0.34.0",
    "claude-agent-sdk>=0.1.0",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.10.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0",
    "types-PyYAML",
]

[project.scripts]
gao-dev = "gao_dev.cli.commands:cli"

[project.urls]
Homepage = "https://github.com/gao-org/gao-dev"
Documentation = "https://github.com/gao-org/gao-dev#readme"
Repository = "https://github.com/gao-org/gao-dev"
Issues = "https://github.com/gao-org/gao-dev/issues"

[tool.setuptools]
packages = ["gao_dev"]

[tool.setuptools.package-data]
gao_dev = [
    "workflows/**/*",
    "agents/*",
    "checklists/*",
    "config/*",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=gao_dev --cov-report=html --cov-report=term"
asyncio_mode = "auto"

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.10.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0",
    "types-PyYAML",
]
